refactor(day-diet): relocate dayDietStore to infrastructure/signals
