+++ src/modules/diet/day-diet/application/usecases/dayChange.ts
@@ -1,4 +1,4 @@
-import { dayDietStore } from '~/modules/diet/day-diet/application/dayDietStore'
+import { dayDietStore } from '~/modules/diet/day-diet/infrastructure/signals/dayDietStore'
 
 export const dayChangeData = dayDietStore.dayChangeData
