+++ src/modules/diet/day-diet/application/usecases/dayState.ts
@@ -1,4 +1,4 @@
-import { dayDietStore } from '~/modules/diet/day-diet/application/dayDietStore'
+import { dayDietStore } from '~/modules/diet/day-diet/infrastructure/signals/dayDietStore'
 
 export const targetDay = dayDietStore.targetDay
 export const setTargetDay = dayDietStore.setTargetDay
